// src/app/maps/interfaces/places.ts
// Generated by https://quicktype.io

export interface PlacesResponse {
  type:        string;
  query:       string[];
  features:    Feature[];
  attribution: string;
}

export interface Feature {
  id:                   string;
  type:                 string;
  place_type:           string[];
  relevance:            number;
  properties:           Properties;
  text_en:              string;
  place_name_en:        string;
  text:                 string;
  place_name:           string;
  matching_place_name?: string;
  center:               number[];
  geometry:             Geometry;
  address?:             string;
  context:              Context[];
}

export interface Context {
  id:           string;
  mapbox_id?:   string;
  text_en:      string;
  text:         string;
  wikidata?:    string;
  language_en?: Language;
  language?:    Language;
  short_code?:  ShortCode;
}

export enum Language {
  En = "en",
}

export enum ShortCode {
  Us = "us",
  UsDc = "US-DC",
  UsWa = "US-WA",
}

export interface Geometry {
  type:          string;
  coordinates:   number[];
  interpolated?: boolean;
  omitted?:      boolean;
}

export interface Properties {
  accuracy:             string;
  "override:postcode"?: string;
  mapbox_id:            string;
}